add search filter controller

+++ src/main/java/com/group_3/kanbanboard/controller/ModelViewTaskController.java
@@ -6,10 +6,7 @@ import com.group_3.kanbanboard.rest.dto.ReleaseResponseDto;
 import com.group_3.kanbanboard.rest.dto.TaskRequestDto;
 import com.group_3.kanbanboard.rest.dto.TaskResponseDto;
 import com.group_3.kanbanboard.rest.dto.UserResponseDto;
-import com.group_3.kanbanboard.service.PrincipalService;
-import com.group_3.kanbanboard.service.ProjectService;
-import com.group_3.kanbanboard.service.ReleaseService;
-import com.group_3.kanbanboard.service.TaskService;
+import com.group_3.kanbanboard.service.*;
 import com.group_3.kanbanboard.service.impl.ModelViewProjectService;
 import com.group_3.kanbanboard.service.impl.ModelViewTaskService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,6 +22,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/projects/{projectId}/releases/{releaseId}/tasks")
@@ -35,17 +33,23 @@ public class ModelViewTaskController {
     private final PrincipalService principalService;
     private final TaskService taskService;
     private final ReleaseService releaseService;
+    private final UtilService utilService;
 
 
     @Autowired
     public ModelViewTaskController(ModelViewTaskService modelViewTaskService,
-                                   ModelViewProjectService modelViewProjectService, PrincipalService principalService,
-                                   TaskService taskService, ProjectService projectService, ReleaseService releaseService) {
+                                   ModelViewProjectService modelViewProjectService,
+                                   PrincipalService principalService,
+                                   TaskService taskService,
+                                   ProjectService projectService,
+                                   ReleaseService releaseService,
+                                   UtilService utilService) {
         this.modelViewTaskService = modelViewTaskService;
         this.modelViewProjectService = modelViewProjectService;
         this.principalService = principalService;
         this.taskService = taskService;
         this.releaseService = releaseService;
+        this.utilService = utilService;
     }
 
 
@@ -60,7 +64,25 @@ public class ModelViewTaskController {
         ReleaseResponseDto releaseResponseDto = releaseService.getById(releaseId);
         model.addAttribute("release", releaseResponseDto);
 
-        return "taskList";
+        return "tasks/taskList";
+    }
+
+    @GetMapping(params = "search")
+    public String searchTasks(@PathVariable UUID projectId,
+                              @PathVariable UUID releaseId,
+                              @RequestParam String search,
+                              Model model){
+
+        List<TaskResponseDto> taskResponseDtoList =
+                modelViewTaskService.getTasksFromProjectAndRelease(projectId, releaseId);
+
+      List<TaskResponseDto> filterList = taskResponseDtoList.stream().filter( task -> task.getTitle().contains(search)).collect(Collectors.toList());
+        model.addAttribute("tasksList", filterList);
+
+        ReleaseResponseDto releaseResponseDto = releaseService.getById(releaseId);
+        model.addAttribute("release", releaseResponseDto);
+
+        return "tasks/taskList";
     }
 
 
@@ -84,12 +106,15 @@ public class ModelViewTaskController {
         List<UserResponseDto> projectUsers = modelViewProjectService.getUsersForProject(projectId);
         model.addAttribute("projectUsers", projectUsers);
 
-        return "taskDetail";
+        return "tasks/taskDetail";
     }
 
     @DeleteMapping("/{taskId}")
-    public String deleteTask(@PathVariable UUID taskId) {
+    public String deleteTask(@PathVariable UUID projectId,@PathVariable UUID taskId) {
+
+        utilService.checkLeadAccess(projectId);
         taskService.deleteTask(taskId);
+
         return "redirect:/projects/{projectId}/releases/{releaseId}/tasks";
     }
 
@@ -101,6 +126,8 @@ public class ModelViewTaskController {
                              @ModelAttribute TaskRequestDto taskRequestDto,
                              Model model) {
 
+        utilService.checkLeadAccess(projectId);
+
         TaskResponseDto distinctTask =
                 modelViewTaskService.setDependenciesAndSave(taskId, projectUserSelect, projectId, releaseId, taskRequestDto);
         model.addAttribute("distinctTask", distinctTask);
@@ -111,7 +138,7 @@ public class ModelViewTaskController {
         List<UserResponseDto> projectUsers = modelViewProjectService.getUsersForProject(projectId);
         model.addAttribute("projectUsers", projectUsers);
 
-        return "taskDetail";
+        return "tasks/taskDetail";
     }
 
     @GetMapping("/addTask")
@@ -135,7 +162,7 @@ public class ModelViewTaskController {
         List<UserResponseDto> projectUsers = modelViewProjectService.getUsersForProject(projectId);
         model.addAttribute("projectUsers", projectUsers);
 
-        return "taskDetail";
+        return "tasks/taskDetail";
     }
 
     @PostMapping
@@ -144,6 +171,9 @@ public class ModelViewTaskController {
                           @RequestParam String projectUserSelect,
                           @ModelAttribute TaskRequestDto taskRequestDto,
                           Model model){
+
+        utilService.checkLeadAccess(projectId);
+
         TaskResponseDto distinctTask = modelViewTaskService
                 .setDependenciesAndSave(projectUserSelect, projectId, releaseId, taskRequestDto);
